Refactor: 자료형 변경으로인한 코드수정

+++ grid/src/main/java/org/highfives/grid/vacation/command/entity/VacationInfo.java
@@ -15,7 +15,7 @@ public class VacationInfo {
     private int id;
 
     @Column(nullable = false, name = "vacation_num")
-    private int vacationNum;
+    private double vacationNum;
 
     @Column(nullable = false, name = "add_time")
     private String addTime;
